+++ src/game_state.rs
@@ -1,0 +1,33 @@
+use crate::house::*;
+
+// https://doc.rust-lang.org/book/ch10-02-traits.html
+pub trait Observable {
+    fn summarize(&self, game_state: &GameState) -> String;
+}
+
+pub enum GameStateName {
+    House,
+}
+
+pub struct GameState {
+    pub house: House,
+    pub current_location: GameStateName
+}
+
+impl GameState {
+    pub fn new() -> Self {
+        GameState { 
+            house: House::new(),
+            current_location: GameStateName::House}
+    }
+
+    pub fn summarize(&self) -> String {
+        self.get_current_location().summarize(&self)
+    }
+
+    pub fn get_current_location(&self) -> &impl Observable {
+        match self.current_location {
+            GameStateName::House => &self.house
+        }
+    }
+}