+++ src/main.rs
@@ -1,7 +1,8 @@
-mod locations;
-use locations::*;
+mod game_state;
+mod house;
+use game_state::*;
 
 fn main() {
-    let house = House::new();
-    println!("{}", house.summarize());
+    let game = GameState::new();
+    println!("{}", game.summarize());
 }